Move VL routing into MoE layers
