--- python/sglang/srt/multimodal/dsv41/__init__.py
@@ -1,1 +1,0 @@
-"""Vision tower, image preprocessing and VL expert routing for DeepSeek-V4.1."""
