+++ test/registered/kernels/ops/moe/test_dsv41_vl_router.py
@@ -5,7 +5,7 @@ from types import SimpleNamespace
 import torch
 import torch.nn.functional as F
 
-from sglang.srt.multimodal.dsv41.vl_routing import vision_topk
+from sglang.srt.layers.moe.vision_topk import vision_topk
 from sglang.test.ci.ci_register import register_cuda_ci
 
 register_cuda_ci(est_time=30, stage="base-b", runner_config="1-gpu-small")
